move Optimizer class to its own file
